Allowing rules to return a number of validation messages, instead of just one
