+++ AutoRest/Modelers/Swagger/Validators/Attributes/RuleAttribute.cs
@@ -26,60 +26,12 @@ namespace Microsoft.Rest.Modeler.Swagger.Validators
         {
             if (Rule != null)
             {
-                object[] outParams;
-                if (!Rule.IsValid(obj, out outParams))
+                foreach(var message in Rule.GetValidationMessages(obj))
                 {
-                    yield return CreateException(null, Rule.Exception, outParams);
+                    yield return message;
                 }
             }
             yield break;
         }
-
-        protected ValidationMessage CreateException(SourceContext source, ValidationException exceptionId, params object[] messageValues)
-        {
-            ValidationMessage validationMessage;
-            ValidationException[] ignore = new ValidationException[] { };
-            if (ignore.Any(id => id == exceptionId))
-            {
-                validationMessage = new ValidationMessage()
-                {
-                    Severity = LogEntrySeverity.Info,
-                    Message = ""
-                };
-            }
-            else if (ValidationExceptionConstants.Info.Messages.ContainsKey(exceptionId))
-            {
-                validationMessage = new ValidationMessage()
-                {
-                    Severity = LogEntrySeverity.Info,
-                    Message = string.Format(CultureInfo.InvariantCulture, ValidationExceptionConstants.Info.Messages[exceptionId], messageValues)
-                };
-            }
-            else if (ValidationExceptionConstants.Warnings.Messages.ContainsKey(exceptionId))
-            {
-                validationMessage = new ValidationMessage()
-                {
-                    Severity = LogEntrySeverity.Warning,
-                    Message = string.Format(CultureInfo.InvariantCulture, ValidationExceptionConstants.Warnings.Messages[exceptionId], messageValues)
-                };
-            }
-            else if (ValidationExceptionConstants.Errors.Messages.ContainsKey(exceptionId))
-            {
-                validationMessage = new ValidationMessage()
-                {
-                    Severity = LogEntrySeverity.Error,
-                    Message = string.Format(CultureInfo.InvariantCulture, ValidationExceptionConstants.Errors.Messages[exceptionId], messageValues)
-                };
-            }
-            else
-            {
-                throw new NotImplementedException();
-            }
-
-            validationMessage.Source = source;
-            validationMessage.ValidationException = exceptionId;
-            return validationMessage;
-        }
-
     }
 }
